fix: support WASM.
